refine tree contexts; slice context -> lol context
The old term "slice context" was too overloaded and imprecise;
we will now talk about slicing only when discussing multi-argument
subscript operations.  Various confused levels of treeness may now be
distinguished with the .tree method, and the list-of-lists type that
used to be "Slicel" is now "LoL".  The old "slice" contextualizer
is now "lol", and is more or less equivalent to .tree(*.list).LoL,
that is do 1 level of tree mapping to list.

Note these are now lists of List, not of Seq.  That's the only
substantive change (along with the .tree method).  Almost everything
else is just refining the terminology.