[S02] be more explicit about iterating sets/bags
The intent has always been that when you use a set or bag as a list,
it behaves as a list of its keys, regardless of any underlying hash
interface it might also respond to.  You must use .pairs explicitly
to get the hash pairs out of a set or bag as a list.