dispatch generation should work under MI
Fixed an ambiguity noted by sorear++.  It's not just the presence
of multi declarations in the current scope that cause autogeneration
of a dispatch routine.  Any difference in the candidate set at the point
of call should make a fresh dispatcher.  We can reuse parent dispatchers
only if they represent the same set of routines.