don't include positional names in API
As suggested by masak++ and jnthn++, remove the names of positionals
from the public API.

In its place, substitute the more useful ability to treat positionals
as if they were named by the use of *numeric* keys.  That is, move the
ability to name positionals out of the binder and into the coercion
from hash to capture.  This answers the three objections to named positionals:
    * The positions of the positionals are already part of the public API.
    * This has no run-time overhead (unless you force this coercion to happen
	at run time, using prefix:<|> for instance).
    * People will use this because many external formats do not support captures.