gather/take is lexotic with dynamic fallback
We have to allow for implementations and libraries to use gather/take
internally.  To preserve the sanity of the user in such cases,
we need gather/take to behave lexotically, so that if you use take
in the lexical scope of a gather, it always goes to *that* gather,
even if the take is inside a closure that is passed into the scope
of another gather.