make MAIN more robust and less magic
The problem with the previous approach was that it made the parsing
of command lines dependent on the signature of each multi MAIN
candidate, making it impossible to first process @*ARGS and then
do a multi dispatch.

A different approach at fixing that would be to use the signature of
the proto of MAIN as reference instead. If somebody feels strongly
about preserving the old magic, please revert this commit and clarify
that magic command line parsing needs an explicit proto MAIN.