old <a>**',' is now written <a>+%','
The separator syntax is now required to be a modifier on a quantifier,
so the old <a> ** ',' is now written <a>+ % ','.  This allows for other
quantifiers like * or ** on the left as the basis of the constraint,
while disambiguating use of ** to only work as a quantifier without
overloading separator matching.  The C<%> because it can be pronounced
'modulo' and because it looks like the relationship of two things.  And
also because it's unlikely to be confused with other regex forms.