define %%; clarify that % is not quantmod
ingy++ requested a method of stopping optionally after the separator,
so that's now %% (with some notion of "evenly divisible" in some kind
of loose sense).  (Note, if you want to require the separator at the
end, don't use this construct, just use [x ',']* or such.

sorear++ requested clarification: % (and %%) are not really quantifier
modifiers; they modify a quantified atom to insert requirements
between each iteration of that quantified atom.  Hence any quantifier
may still take an additional control quantmod in addition to a separator.