[S26] lower-case $=POD, =DATA and =END
This has the effect that =END and =DATA no longer have to be special
cases to the rule "All uppercase block typenames are reserved for
specifying standard documentation, publishing, or source components."

This commit also renames =END to =finish, because =end would clash
with the end marker of delimited blocks.

Commit prepared by Damian Conway.