[S13] unspec multiple signatures
S06 is fairly stable at this point, and the suggestion in S13 to
support multiple signatures for one single routine never made it
into S06. Neither of three major implementations support multiple
signatures. There's no sense of a general demand for them. The
feature is probably implementable, but the torment to the
implementors isn't proportionate to the benefit for the users.