Failed matches to return Nil, not (Match)
A Match now only ever represents a successful match.  A failed match is
represented by Nil.  Also, we add Match and Nil to the list of things
that are explicitly not allowed as literal tokens to smartmatchers due
to their use as success/failure from .ACCEPTS methods.

And a great sigh of relief comes from those who didn't want the OKness
proposal to go much farther than that...  :-)