more set/bag list clarifications
We're not trying to override explicit @ or .keys here, merely the
implicit policy for a default of item in list context.

Also, set/bag constructors are simply slurpies, and shouldn't care at all
what was passed as the objects in those slurpies.  Generally one should
be using "union" to combine sets or bags anyway, not relying on implicit
constructor semantics to do that.