simplify blessing
The .bless method no longer takes a first candidate argument.
(Saying how to create the candidate is now the responsibility of the
representation.)  The BUILDALL method is now gone, so that the compiler
is free to write a bless submethod that inlines all the creation and
build logic into an easily inline-able routine.  We will probably
invent ways to restore the flexibility of the old system as we go on,
but there's no need to inflict that flexibility on all objects with
well-known representations, nor on the users trying to create such
objects.  That this will also make things faster is not undesirable...