Loops in a statementlist are forced to be eager
Loops used at the top-level of a block where multiple semicolon-separated
statements are expected now are always forced to be eagerly evaluated, even
if they return a value.

Loops used where a single statement is parsed retain lazy semantics.
For instance, inside brackets, or after a blorst of the st persuasion.
Some of the list comprehensions we've written may need parens around
them, or "lazy" in front of them, to retain their current lazy semantics.