Conjectures, s/keeper/vow/ to stay non-commital
Speculations in some spots that could have better syntax.

The word "keeper" kind of implies you're going to keep the promise,
when in fact you might break it.  To avoid this overtone, as well
as other unfortunate meanings of "keeper" ("Yep, he's a keeper!"),
we rename the internal stewardship of the promise to "vow".