Adopt quietfanatic++'s suggestion of Less/More
The problem is not the sign of the numeric values, which is consistent with
common usage in C and Perl.  The problem is that the names seemed to be
answering the opposite question from the "less, same, or greater" operators,
which are asking about the left operand, not the right one.

fixes 61