tweak .join to .finish (to complement .start)
Avoiding unfortunate name collision with Str.join in the usersphere.
Since we're trying to preserve OO notions, we tell the thread to finish
(implying we'll wait for them) rather than use a method name like 'await',
which makes no sense as a message to a thread object, since waiting is
not what the thread is supposed to do.