[S32::Temporal] Change added Duration info to Interval.
Yes, we just made a brand new object. Some new issuess, including the
<start>/<end> form of ISO 8601 intervals, Duration <-> Interval
conversion (potentially), and describing Interval in other locations
(such as S02) need to be taken care of. But it's a bit late for that for
me :) .

Also, Instant and Duration still need documenting in S32::Temporal, and
what to do with the <start>/<duration> and <duration>/<end> interval
forms (if anything).