Rename META.info -> META6.json
Andreas Koenig pointed out that .info is being used for quite some other things
on the Open Source world.  So we (FROGGS + lizmat) decided to change the name
in such a way, that it is more in line with the world and still an easy way to
distinguish a Perl 6 distribution from other distributions.