Use space instead of underscore in Wikipedia reference.
While both forms, `Mu_(negative)` and `Mu (negative)` are correct, only
the form with the space inside is intended to be public version
(however, the version with underscore is also correct because of
technical details - otherwise it would have to represented with `%20`
which is just ugly).

If you disagree with this change, then I propose to change the Perl 6's
name to Perl_6. After all, this is what technically it is on Wikipedia.
Or Perl-6, after all Perl 6 prefers hyphens instead of underscores.

The alternative options are "無 (negative)", "Moo", "Mu.new", "MU",
"Star Trek [Ii]nto Darkness", and some more. However, I'm not going to
consider them, because it would ridiculous.

Then again, this commit changes one character. This is ridiculous
anyway. But it's not that I'm going to consider "Moo" instead of "Mu",
even considering this minor change with huge commit description (yes,
I noticed).