Spec "outcome/whenever" instead of "winner/done/more"
For discussion, and for a test implementation.  Please comment / fire on it /
shoot it down if you think you should.  By no means do I feel this is the
final thing.  It's just that winner {} currently is *so* broken, we either
need to fix that, or come up with something else.  This is just my something
else :-)