Re-purpose the method ^foo(...) { ... } syntax.
The existing design of it was out of line with the way Perl 6 evolved
since then, and also not especially useful. This takes the syntax and
enables its use for per-type meta-behavior specialization. This, as a
special case, allows attaching extra metadata to the meta-object,
which captures some aspect of what the syntax was originally specified
to do. However, it maintains the distinction between objects and their
meta-objects more strongly, which also fits better with the overall
Perl 6 design.