Escaping `<` and `>` inside `C<>` block
This is because the `V<` part of `ENV<PATH>` was being confused with a
non-existent command sequence `V<`.  Since the text should be printed out in
a code block, escaping the internal pointy brackets solves the problem and
makes the text only marginally more difficult (in the POD) to read.  An
alternative would have been to use `Z<>`, however this is less well known
and harder to read in the plain text.