Introducing nl-reading/nl-saying, comments welcome
Earlier versions of the spec and the implementation used only the rather
verbose "input-line-separator".  This was later shortened to "nl", but we
lost the inputness of the separator.  Since we only use the new line separator
on "say", I thought it would be approriate to name that case "nl-saying".
Consequently, extending "nl" to "nl-reading" seemed appropriate.