Update S32::Str to align better with S15
This commit removes quite a bit of cruft from the pre-S15 days, and
additonally reformats the code blocks so that they stand out more in the
file, and for longer function signatures, makes them less crowded.

The only functional changes in this commit is the removal of the special
exception for C<&tcuc> in the C<wordcase> method (since C<&tcuc>
apparently doesn't exist), and a clarification for sprintf that the '%'
directive can only appear as a literal '%%' (both bringing it in line
with C's printf behavior and ensuring nonsense like '%1$+*v0.5%' isn't
accepted).

Ideally this synopsis would talk about the role(s) shared by the string
types, instead of focusing on C<Str> as a model for other types.
However the nature of these potential roles isn't clear enough yet to
make that kind of change to this.