Remove :T smiley from specs
The :U smiley has taken the place of :T over the years, and the only
difference specified between them (where :U accepts Failures but :T does
not) doesn't appear to make sense anymore. Now :U is spec'd to mean type
objects, and :D means instances.

The only potentially significant change for rakudo is the change from :T
to :U as the default smiley for method new's invocant, but depending on
existing implementation may already be how it works.