Add postfix:<i> to S03
It wasn't mentioned at all before, and since it's at the methodcall
level in rakudo, which may not be expected at first, I figured it was
worth mentioning. It also makes postfix:<i> the only operator at that
level to not start with a dot, which seems like another noteworthy
detail.