Update dependency specifications in S22
An attempt at implementation has shows that using lists for dependency
alternatives causes confusion and is not self-documenting enough as the depends
section itself consists of a list. Using an object with an any key is more
clear and lends itself to future expansion.

Collapsing of system specific values has always been meant to also cover
dependencies. This is now stated explicitly and the details on what's available
described.

The build-depends and test-depends keys are deprecated. They should not have
been part of the v1 META spec. Leaving them in was just an oversight.